0.0.6.230612a
1, 修改为使用轮循方式

+++ include/xq/net/timer.hpp
@@ -55,7 +55,7 @@ public:
 
 
     inline void create_timer_at(uint64_t expire_us, T* ev) {
-        if (expire_us <= sys_clock()) {
+        if (expire_us <= (uint64_t)sys_clock()) {
             ev->update();
             if (ev->destory()) {
                 delete ev;
